==> notebooks/02_clean-grid-disruption-00-14-standardized-data.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('../data/Grid_Disruption_00_14_standardized.csv')
data.head()

data.shape

### Drop Canada rows

data = data[~((data['Geographic Areas']!= np.nan) & (data['Geographic Areas'].str.contains('Canada')))]

data['Number of Customers Affected'].isnull().sum()

data['Number of Customers Affected'].value_counts()

### Convert 'Uknown' and like strings to Nan

data['Number of Customers Affected'] = data['Number of Customers Affected'].map(lambda x: x if (type(x) == float) else np.nan if (x.isalpha() ==True) else x)

### Convert '-' to Nan

data['Number of Customers Affected'] = data['Number of Customers Affected'].map(lambda x: x if (type(x) == float) else np.nan if (x=='-') else x)


data['Number of Customers Affected'].isnull().sum()

### Remove commas

data['Number of Customers Affected'] = data['Number of Customers Affected'].map(lambda x: x.replace(',','') if type(x) != float else x)


### Remove 'Approx. '

data['Number of Customers Affected'] = data['Number of Customers Affected'].map(lambda x: x.replace('Approx. ','')if type(x) != float else x)


data[(~data['Number of Customers Affected'].isnull())&data['Number of Customers Affected'].str.contains(' ')]

### Manually set specific rows

data.loc[1637,'Number of Customers Affected']= 2000000    # 2 million
data.loc[1191,'Number of Customers Affected']= 11175      # GRE (1900) Total 11175
data.loc[1262,'Number of Customers Affected']= 111000     # 111000 (peak)
data.loc[1290,'Number of Customers Affected']= np.nan     # Major Industrial Customer Load Reduction
data.loc[1299,'Number of Customers Affected']= 4          # Interruptible Tarriff 1-6 customers
data.loc[1301,'Number of Customers Affected']= 2500000    # 700000 (peak) 2500000 (actual)
data.loc[1323,'Number of Customers Affected']= 9000       # 8000 -10000
data.loc[1351,'Number of Customers Affected']= np.nan     # --
data.loc[1373,'Number of Customers Affected']= 1100000    # 1.1 million and 100000 gas customers
data.loc[1383,'Number of Customers Affected']= np.nan     # PG&E
data.loc[1412,'Number of Customers Affected']= np.nan     # PG&E
data.loc[1496,'Number of Customers Affected']= 250000     # Greater than 250000
data.loc[1498,'Number of Customers Affected']= 225353     # 94366 at peak total 225353
data.loc[1501,'Number of Customers Affected']= 62500      # 62500 at peak
data.loc[1505,'Number of Customers Affected']= 85000      # 85000 at peak
data.loc[1512,'Number of Customers Affected']= np.nan     # PG&E
data.loc[1529,'Number of Customers Affected']= 6          # 6 (utilities)
data.loc[1530,'Number of Customers Affected']= 200000     # 200000 (Peak)
data.loc[1532,'Number of Customers Affected']= 104195     # 104195 at 5:23 p.m. 11/13/03
data.loc[1540,'Number of Customers Affected']= 108000     # 108000 (Dist. And Trans. Combined)
data.loc[1541,'Number of Customers Affected']= 530000     # Over 530000 peak on 9/19/03
data.loc[1543,'Number of Customers Affected']= 50000      # Under 50000
data.loc[1545,'Number of Customers Affected']= 32000      # peak 32000 9/18/03 7:00 p.m.
data.loc[1547,'Number of Customers Affected']= 1800000    # 1.8 million
data.loc[1548,'Number of Customers Affected']= 4          # 4 (industrial)
data.loc[1550,'Number of Customers Affected']= 133000     # 93000 at peak 133000 cumulative
data.loc[1576,'Number of Customers Affected']= 115739     # 102842 (Georgia); 12897 (Alabama)
data.loc[1582,'Number of Customers Affected']= 340000     # over 340000
data.loc[1587,'Number of Customers Affected']= 1500000    # 1.5 million
data.loc[1595,'Number of Customers Affected']= 1          # 1 PG&E
data.loc[1599,'Number of Customers Affected']= 1          # 1 PG&E
data.loc[1600,'Number of Customers Affected']= 1          # 1 PG&E
data.loc[1609,'Number of Customers Affected']= 36073      # (residential) 36073
data.loc[1612,'Number of Customers Affected']= 127566     # 70848; 56718
data.loc[1615,'Number of Customers Affected']= 50000      # 25000 per hour
data.loc[1645,'Number of Customers Affected']= 425000     # 40000-45000

data['Number of Customers Affected'] =data['Number of Customers Affected'].map(lambda x: int(x)if type(x) != float else x)


data.to_csv('../data/Grid_Disruption_00_14_standardized_cleaned.csv', index=False)
