==> notebooks/03_a_preprocessing.ipynb <==
# Pull cities mentioned from text
> Author: Sharnique Beck

Packages used:
- Python 3.6.6
- pandas 0.23.4

Pull cities mentioned in tweet text into a list to identify location of poweroutage.


import pandas as pd
import pickle
import re

# read in dictionary of cities
cities = pickle.load( open( "../data/cities.p", "rb" ) )
cities.keys()

# read in data frame
df = pd.read_csv('../data/combined_tweets_outages.csv' )

df.head()

df['text'][17]

df['city_pull']=''

count = 0
for row in range(df.shape[0]):                                # loop through rows of df
    row_lst = []                                              # create empty list for cities pulled
    text = re.findall('([\w]+)\s', df['text'][row])          # find all words
    added = re.findall('[@|#]([A-z]+)', df['text'][row])      # find all mentions and hashtags
    text.extend(added)                                        # list of words in 'text' string
    for word in text:                                         # loop through each word in 'text' string
        for state,city in cities.items():
            for t_city in city:
                for name in t_city:
                     if word == name:
                        row_lst.append(str(name+', '+state))
        if len(row_lst) !=0:
            df['city_pull'][row] = row_lst                    # add list to df if list has values
        else:
            df['city_pull'][row] = ''                         # else add empty string to df
    count +=1
    if count % 100 == 0:
        print(count)

df

df.to_csv('../data/combined_tweets_outages_cities.csv', index=False)
