==> notebooks/03_b_preprocessing.ipynb <==
# Pull states mentioned in text and user names
> Author: Sharnique Beck

Packages used:
- Python 3.6.6
- pandas 0.23.4

This notebook begins the process of pulling state abbreviations from the tweet text and usernames. This can then be used to verify cities pulled with states mentioned that have multiple states containg the same city names.

import pandas as pd
import regex as re


# import data
df = pd.read_csv('../data/combined_tweets_outages_cities.csv' )

# Drop columns not used
df.drop(columns = ['likes','replies','retweets'],  axis = 1, inplace=True)

df.head()

df['city_pull'] = df['city_pull'].map(lambda x: '' if type(x)!=str else x)

### Pull state abbreviation from text

state =['MA','ME','NY','RI','VT','CT','NH'] 

# Look for state abbreviations mentioned in text column
df['text_state_pull'] = df['text'].map(lambda x: re.findall(r'([A-Z]{2})',x)) 
df['text_state_pull'] = df['text_state_pull'].map(lambda lst: [st for st in lst if st in state])
df['text_state_pull'] = df['text_state_pull'].map(lambda x: '' if x==[] else x[0])


### Pull state abbreviation from username

# Look for state abbreviations located in user names
df['user_state_pull']=df['user'].map(lambda x: re.findall(r'([A-Z]{2})',x)) 

df['user_state_pull'] = df['user_state_pull'].map(lambda lst: [st for st in lst if st in state])
df['user_state_pull'] = df['user_state_pull'].map(lambda x: '' if x==[] else x[0])


df

df.to_csv('../data/combined_tweets_outages_cities_states.csv',index=False)
