==> notebooks/03_preprocessing.ipynb <==
# Preprocessing
> Author: [Dawn Graham](https://dawngraham.github.io/)

Quick clean of `monthlytweets.csv`.

Versions used:
- Python 3.6.6
- pandas 0.23.4

### Import libraries

import pandas as pd

### Read in data

tweets = pd.read_csv('../data/raw-data/monthlytweets.csv')

### Initial cleaning

# Drop duplicates
tweets.drop_duplicates(inplace=True)
print(f"Total: {tweets.shape[0]}")
print(f"Unique: {tweets['id'].nunique()}")

sorted(tweets['timestamp'].unique(), reverse=True)[:3]

# # Get number of observations containing word
tweets[tweets['timestamp'].str.contains('http://bit.ly/yTIrUE\xa0', case=False, regex=True)]

# Drop row 7454
tweets.drop(tweets.index[7454], inplace=True)

# Drop rows where timestamp == timestamp
tweets = tweets[tweets['timestamp'] != 'timestamp']

#Drop rows where query is null
tweets = tweets[tweets['query'].isnull() != True]

### Check shape & types

tweets.shape

tweets.dtypes

tweets.isnull().sum()

### Change to timeseries

# Set `timestamp` to datetime and set it to index
tweets['timestamp'] = pd.to_datetime(tweets['timestamp'])
tweets.set_index('timestamp', inplace=True)
tweets.head()

### Export as `monthlytweets_cleaned.csv`

tweets.to_csv('../data/monthlytweets_cleaned.csv', index=True)
