==> notebooks/06_outages_state_and_time.ipynb <==
# Power outages with state & time info
> Author: [Dawn Graham](https://dawngraham.github.io/)

The first part of this notebook assigns location of power outages (as state abbreviations) to power outages. It then transforms event and restoration info to datetime values.

The last adds a column to tweets data indicating if the tweet was at the same time as a verified outage.

Versions used:
- Python 3.6.6
- pandas 0.23.4
- numpy 1.15.4

### Import libraries

import pandas as pd
import numpy as np
from collections import OrderedDict

### Get data

df = pd.read_csv('../data/Grid_Disruption_00_14_standardized_cleaned.csv')
df.head()

df.shape

### Drop rows
Drop rows with missing geographic data.

# # Look at unique Geographic Areas
# sorted(df['Geographic Areas'].astype(str).unique())

df = df[df['Geographic Areas'] != '-']
df = df[df['Geographic Areas'].notnull()]

### State abbreviations dictionary
Bring in dictionary of state names and abbreviations. From [mGalarnyk/us_state_abbrev.py](https://gist.github.com/mGalarnyk/27c99c5f55133a2ceb42b8c9a450d794).

us_state_abbrev = {
    'Alabama': 'AL',
    'Alaska': 'AK',
    'Arizona': 'AZ',
    'Arkansas': 'AR',
    'California': 'CA',
    'Colorado': 'CO',
    'Connecticut': 'CT',
    'Delaware': 'DE',
    'Florida': 'FL',
    'Georgia': 'GA',
    'Hawaii': 'HI',
    'Idaho': 'ID',
    'Illinois': 'IL',
    'Indiana': 'IN',
    'Iowa': 'IA',
    'Kansas': 'KS',
    'Kentucky': 'KY',
    'Louisiana': 'LA',
    'Maine': 'ME',
    'Maryland': 'MD',
    'Massachusetts': 'MA',
    'Michigan': 'MI',
    'Minnesota': 'MN',
    'Mississippi': 'MS',
    'Missouri': 'MO',
    'Montana': 'MT',
    'Nebraska': 'NE',
    'Nevada': 'NV',
    'New Hampshire': 'NH',
    'New Jersey': 'NJ',
    'New Mexico': 'NM',
    'New York': 'NY',
    'North Carolina': 'NC',
    'North Dakota': 'ND',
    'Ohio': 'OH',
    'Oklahoma': 'OK',
    'Oregon': 'OR',
    'Pennsylvania': 'PA',
    'Rhode Island': 'RI',
    'South Carolina': 'SC',
    'South Dakota': 'SD',
    'Tennessee': 'TN',
    'Texas': 'TX',
    'Utah': 'UT',
    'Vermont': 'VT',
    'Virginia': 'VA',
    'Washington': 'WA',
    'West Virginia': 'WV',
    'Wisconsin': 'WI',
    'Wyoming': 'WY',
    'Puerto Rico': 'PR',
}

### Create `State` column

# Set up empty column
df['State'] = ''

If `Geographic Areas` contains state name or abbreviation, include abbreviation in `State` column.

for state in us_state_abbrev:
    df.loc[df['Geographic Areas'].str.contains(state), 'State'] += f'{us_state_abbrev[state]} '
    df.loc[((df['Geographic Areas'].str.contains(us_state_abbrev[state])) & (us_state_abbrev[state] not in df['State'])), 'State'] += f'{us_state_abbrev[state]} '

Fill in missing values in `State` column.

df.loc[(df['Geographic Areas'].str.contains('Bakersfield') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Geographic Areas'].str.contains('Los Angeles') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Geographic Areas'].str.contains('Sacramento') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Geographic Areas'].str.contains('San Diego') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Geographic Areas'].str.contains('San Francisco') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Geographic Areas'].str.contains('Connecicut') & ('CT ' not in df['State'])), 'State'] += f'CT '
df.loc[(df['Geographic Areas'].str.contains('Maui') & ('HI ' not in df['State'])), 'State'] += f'HI '
df.loc[(df['Geographic Areas'].str.contains('Oahu') & ('HI ' not in df['State'])), 'State'] += f'HI '
df.loc[(df['Geographic Areas'].str.contains('Akron') & ('OH ' not in df['State'])), 'State'] += f'OH '
df.loc[(df['Geographic Areas'].str.contains('Cleveland') & ('OH ' not in df['State'])), 'State'] += f'OH '
df.loc[(df['Geographic Areas'].str.contains('Long Island') & ('NY ' not in df['State'])), 'State'] += f'NY '
df.loc[(df['Geographic Areas'].str.contains('Houston') & ('TX ' not in df['State'])), 'State'] += f'TX '

df.loc[(df['Respondent'].str.contains('California') & ('CA ' not in df['State'])), 'State'] += f'CA '
df.loc[(df['Respondent'].str.contains('Florida') & ('FL ' not in df['State'])), 'State'] += f'FL '
df.loc[(df['Respondent'].str.contains('Detroit') & ('MI ' not in df['State'])), 'State'] += f'MI '
df.loc[(df['Respondent'].str.contains('Texas') & ('TX ' not in df['State'])), 'State'] += f'TX '
df.loc[(df['Respondent'].str.contains('Puget Sound') & ('WA ' not in df['State'])), 'State'] += f'WA '
df.loc[(df['Respondent'].str.contains('Puerto Rico') & ('PR ' not in df['State'])), 'State'] += f'PR '

Correct incorrect or missing `State` values.

df.loc[df['Geographic Areas'].str.contains('VA')]
df.loc[1009, 'State'] = 'TN '

df.loc[876, 'State'] = 'CO '
df.loc[1006, 'State'] = 'KY '
df.loc[1073, 'State'] = 'WA '
df.loc[1107, 'State'] = 'CT ME MA NH RI VT '
df.loc[1139, 'State'] = 'FL '
df.loc[1210, 'State'] = 'AZ '

Update rows that included were flagged as `VA` because of West Virginia.

df.loc[12, 'State'] = 'WV '
df.loc[76, 'State'] = 'MD WV '
df.loc[317, 'State'] = 'WV '
df.loc[328, 'State'] = 'MD WV '
df.loc[402, 'State'] = 'MD WV '
df.loc[405, 'State'] = 'IN MI OH WV '
df.loc[410, 'State'] = 'WV '
df.loc[693, 'State'] = 'KY WV '
df.loc[733, 'State'] = 'KY WV '
df.loc[836, 'State'] = 'OH WV KY '
df.loc[836, 'State'] = 'PA WV '
df.loc[918, 'State'] = 'WV KY OH '
df.loc[991, 'State'] = 'KY WV OH '
df.loc[1340, 'State'] = 'MD WV PA '

## Check remaining rows with no `State` value.

df[df['State'] == ''].shape

df[df['State'] == '']

## Get begin / end times

df['Year'].value_counts()

# Quick filter to only use data from 2009 - 2012
df = df[(df['Year'] >= 2009) & (df['Year'] <= 2012)]

# Check to see if any begin/end times are missing
df.isnull().sum()

### Get datetime values

# Combine Began date & time into 'Outage Start' as datetime
df.loc[:,'Outage Start'] = pd.to_datetime(df['Date Event Began'].astype(str)+' '+df['Time Event Began'].astype(str))

# Drop the two rows where 'Date of Restoration' == 'Ongoing'
df = df[df['Date of Restoration'] != 'Ongoing']

# Combine Restoration date & time into 'Outage End' as datetime
df.loc[:,'Outage End'] = pd.to_datetime(df['Date of Restoration'].astype(str)+' '+df['Time of Restoration'].astype(str))

# Save only desired columns to outages.csv
df = df[['Respondent', 'Geographic Areas', 'State', 'Number of Customers Affected', 'Outage Start', 'Outage End']]
df.to_csv('./data/outages.csv')
df.head(3)

---
## Check if tweets coincide with power outages

tweets = pd.read_csv('../data/combined_monthlytweets.csv')

# Set `timestamp` to datetime and set it to index
tweets['timestamp'] = pd.to_datetime(tweets['timestamp'])
tweets.set_index('timestamp', inplace=True)
tweets.head(3)

# Dataframe contains sponsored tweets outside of our time of interest (September 2009 to end of December 2012).
# Filtering to dates of interest.
tweets = tweets[:'2012']

# Set up 'outage' column with all values set to 0 and empty 'outage_state' column
tweets['outage'] = 0
tweets['outage_state'] = ''

tweets.shape

count = 0
for i in range(len(tweets)):
    count += 1
    for row in range(len(df)):
        if ((df.iloc[row]['Outage Start'] <= tweets.index[i]) & (tweets.index[i] <= df.iloc[row]['Outage End'])):
#             ## Print to confirm tweet times are during an outage
#             print('Outage start:', df.iloc[row]['Outage Start'])
#             print('Tweet time:  ', tweets.index[i])
#             print('Outage end:  ', df.iloc[row]['Outage End'])
#             print('Outage state: ', df.iloc[row]['State'])
#             print('')
            tweets.loc[tweets.index[i], 'outage'] = 1
            tweets.loc[tweets.index[i], 'outage_state'] = tweets['outage_state'][i] + ' ' + f"{df.iloc[row]['State']}"
#             break
    if (count) % 100 == 0:
         print(count, end=' ')

# Remove duplicates in `outage state` column
tweets['outage_state'] = tweets['outage_state'].str.split()\
         .apply(lambda x: OrderedDict.fromkeys(x).keys()).str.join(' ')

tweets.head()

tweets['outage'].value_counts()

tweets.to_csv('../data/combined_tweets_outages.csv')
