==> notebooks/08_tweet_vectorizer_clustering.ipynb <==
# Tweet Vectorizer & Clustering
> Author: [Dawn Graham](https://dawngraham.github.io/)

This notebook preprocesses and vectorizes our tweet text so it can be used in models. I then use logistic regression to determine the words most associated with tweets made during power outages.

Versions used:
- Python 3.6.6
- matplotlib 3.0.2
- nltk 3.3
- numpy 1.15.4
- pandas 0.23.4
- regex 2018.11.22
- seaborn 0.9.0
- sklearn 0.0
- Unidecode 1.0.23

## Import libraries

# General
import pandas as pd
import numpy as np

# For natural language processing
import regex as re
import unidecode
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import CountVectorizer

import warnings
warnings.filterwarnings("ignore", 'This pattern has match groups')

# For logistic regression
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix

# For K-Means
from sklearn.cluster import KMeans, k_means
from sklearn.metrics import silhouette_score
from sklearn.datasets.samples_generator import make_blobs

import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib

%matplotlib inline

## Read in data

tweets = pd.read_csv('../data/combined_tweets_outages.csv')

# Set `timestamp` to datetime and set it to index
tweets['timestamp'] = pd.to_datetime(tweets['timestamp'])
tweets.set_index('timestamp', inplace=True)
tweets.head()

## Preprocessing

# Preprocessing function
def text_to_words(raw_text):
    
    # Get rid of accents
    unaccented = unidecode.unidecode(raw_text)
    
    # Get rid of punctuation
    letters_only = re.sub("[^a-zA-Z]", " ", unaccented)
    
    # Get all lowercase words
    words = letters_only.lower().split()
    
    # Instantiate and run Lemmatizer
    lemmatizer = WordNetLemmatizer()
    tokens_lem = [lemmatizer.lemmatize(i) for i in words]
    
    # Remove stop words
    stops = set(stopwords.words('english'))
    meaningful_words = [w for w in words if not w in stops]
    
    # Join into string and return the result.
    return(" ".join(meaningful_words))

# Clean all tweets
total_tweets = tweets.shape[0]
clean_texts = []

print("Cleaning and parsing the tweets...")

j = 0
for text in tweets['text']:
    # Convert to words, then append to clean_tweets
    clean_texts.append(text_to_words(text))
    
    # If the index is divisible by 1000, print a message
    if (j+1) % 1000 == 0:
        print(f'Comment {j+1} of {total_tweets}.')
    
    j += 1
    
    if j == total_tweets:
        print('Done.')

# Add cleaned tweets to dataframe
tweets = tweets.assign(clean_text = clean_texts)
tweets.head(3)

## Train / Test Split

# Set features and target
features = tweets['clean_text']

X = features
y = tweets['outage']

# Train/test split
X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=42)

## Bag of Words

vect = CountVectorizer()

# Fit the vectorizer on our corpus and transform
X_train_vect = vect.fit_transform(X_train)
X_train_vect = pd.DataFrame(X_train_vect.toarray(), columns=vect.get_feature_names())

# Transform the test set
X_test_vect = vect.transform(X_test)

# Transform entire set for K-Means clustering later
X_vect = vect.transform(X)

## Logistic Regression
Use logistic regression to get words that are most likely to be from tweets made during power outages. Cross-validation and accuracy scores help serve as a sanity check.

# Get baseline accuracy score
y_train.value_counts(normalize=True)[1]

# Instantiate model
logreg = LogisticRegression(solver='liblinear')

# Fit on training data.
logreg.fit(X_train_vect, y_train)

# Get scores
print('CV score:', cross_val_score(logreg, X_train_vect, y_train, cv=3).mean())
print('Training accuracy:', logreg.score(X_train_vect, y_train))
print('Testing accuracy:', logreg.score(X_test_vect, y_test))

#### Confusion Matrix
Check confusion matrix to get a sense of how the model is classifying tweets.

# Create confusion matrix
predictions = logreg.predict(X_test_vect)
cm = confusion_matrix(y_test, predictions)
cm_df = pd.DataFrame(cm, columns=['predict neg', 'predict pos'], index=['actual neg', 'actual pos'])
cm_df

## Associated Words

# Create dataframe with coefs and e^coefs for each word
coefs = list(zip(vect.get_feature_names(), logreg.coef_[0].T))
coefs = pd.DataFrame(coefs, columns = ['word','coef'])
coefs['e^coef'] = np.exp(coefs['coef'])

#### Tweets made during power outages
Reminder: Tweets were not necessarily *about* power outages or in locations were outages occurred, but were made at times that there were verified power outages.

# Show words most associated with tweets made during power outages
coefs.sort_values(by='e^coef', ascending=False).head(20)

#### Explore tweets with specific words

# Function to get full tweets and count of tweets containing specific words
def get_tweets(word):
    mask = tweets[tweets['clean_text'].str.contains(f'(^|\W){word}($|\W)')].index
    count = 0
    for i in mask:
        count += 1
        print(i, tweets['text'][i], '\n')
    print(f'\nTotal tweets containing "{word}": {count}')

# Look at titles with words most associated with tweets made during power outages
get_tweets('nhsandy')

#### Notes on findings
- Words in our dataset like `mairene`, `irene`, `nhirene`, `nhsandy`, `risandy`, etc. are highly associated with tweets made during power outages and are usually about outages. While these specific words would not be useful in detecting outages following a future event, they suggest that the name of a given storm (and the name combined with state abbreviations) could be useful in detecting outages following an event.
- Tweets containing `etrs` (standing for Estimated Time of Restorations) could be useful, but most of these are made by power companies providing updates after they already know about outages.
- `ojatg` and `fccenw` are from National Grid's short link to their Outage Central interactive outage map with ETRs. These links are no longer used, apparently because they have shifted to state-specific maps. `Outage Central` and `outage map` could be useful, however are similar to `etrs` and may not provide info about outage locations that power companies don't already know about.
- Taking a closer look at the tweets in our dataset `earthday` and `presidents` suggests these won't be helpful words to use as indicators of power outages.

## K-Means Clustering

kmeans = KMeans(n_clusters=3)
model = kmeans.fit(X_vect)

# Attach predicted cluster to `tweets` dataframe
tweets['predictions'] = model.labels_
tweets.head(3)

tweets.loc[tweets['predictions'] == 0, 'text']

tweets.loc[tweets['predictions'] == 1, 'text']

tweets.loc[tweets['predictions'] == 2, 'text']

## Create a dataframe for cluster_centers (centroids)
centroids = pd.DataFrame(model.cluster_centers_)
centroids.head()
